use mimeTypeToVideoCodecString
